requirements.txt
accelerate
colored
controlnet_aux==0.0.6
cuda-python
diffusers==0.23.1
ftfy
matplotlib
nvtx
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
scipy
transformers==4.31.0
--extra-index-url https://pypi.ngc.nvidia.com
onnx-graphsurgeon>=0.3.27
polygraphy
fastapi
uvicorn
torch
torchvision